CTAA-1614 - Moved exposure state from DashboardViewModel to ExposureNotificationManager
