+++ app/src/main/java/at/roteskreuz/stopcorona/di/RepositoryModule.kt
@@ -1,6 +1,8 @@
 package at.roteskreuz.stopcorona.di
 
 import android.bluetooth.BluetoothAdapter
+import at.roteskreuz.stopcorona.model.managers.ExposureNotificationManager
+import at.roteskreuz.stopcorona.model.managers.ExposureNotificationManagerImpl
 import at.roteskreuz.stopcorona.model.repositories.*
 import org.koin.dsl.module.module
 
@@ -92,6 +94,17 @@ val repositoryModule = module {
             appDispatchers = get(),
             bluetoothManager = get(),
             exposureNotificationClient = get()
+        )
+    }
+
+    single<ExposureNotificationManager> {
+        ExposureNotificationManagerImpl(
+            appDispatchers = get(),
+            preferences = get(),
+            exposureNotificationRepository = get(),
+            bluetoothRepository = get(),
+            googlePlayAvailability = get(),
+            contextInteractor = get()
         )
     }
 }
