+++ app/src/main/java/at/roteskreuz/stopcorona/di/ViewModelModule.kt
@@ -50,15 +50,12 @@ val viewModelModule = module {
         DashboardViewModel(
             appDispatchers = get(),
             dashboardRepository = get(),
-            contextInteractor = get(),
             infectionMessengerRepository = get(),
             quarantineRepository = get(),
             configurationRepository = get(),
-            exposureNotificationRepository = get(),
             databaseCleanupManager = get(),
-            googlePlayAvailability = get(),
             changelogManager = get(),
-            bluetoothRepository = get()
+            exposureNotificationManager = get()
         )
     }
